+++ Reference Implementation/OANDARestLibrary/TradeLibrary/DataTypes/Communications/DeletePositionResponse.cs
@@ -8,7 +8,7 @@ namespace OANDARestLibrary.TradeLibrary.DataTypes.Communications
 {
 	public class DeletePositionResponse : Response
 	{
-		public List<int> ids { get; set; }
+		public List<long> ids { get; set; }
 		public string instrument { get; set; }
 		public int totalUnits { get; set; }
 		public double price { get; set; }
